refactor: moment, comment service 계층으로 변경
