+++ server/src/main/java/moment/comment/service/CommentImageService.java
@@ -1,4 +1,4 @@
-package moment.comment.application;
+package moment.comment.service;
 
 import java.util.HashMap;
 import java.util.List;
