+++ server/src/main/java/moment/comment/service/CommentQueryService.java
@@ -1,4 +1,4 @@
-package moment.comment.application;
+package moment.comment.service;
 
 import java.util.List;
 import moment.comment.domain.Comment;
