+++ server/src/main/java/moment/comment/service/DefaultCommentQueryService.java
@@ -1,4 +1,4 @@
-package moment.comment.application;
+package moment.comment.service;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
