+++ server/src/main/java/moment/moment/service/MomentImageService.java
@@ -1,4 +1,4 @@
-package moment.moment.application;
+package moment.moment.service;
 
 import java.util.HashMap;
 import java.util.List;
