+++ server/src/main/java/moment/moment/service/MomentTagQueryService.java
@@ -1,4 +1,4 @@
-package moment.moment.application;
+package moment.moment.service;
 
 import java.util.List;
 import moment.moment.domain.Moment;
