+++ server/src/main/java/moment/moment/service/TagService.java
@@ -1,4 +1,4 @@
-package moment.moment.application;
+package moment.moment.service;
 
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
