+++ server/src/test/java/moment/moment/service/MomentImageServiceTest.java
@@ -1,4 +1,4 @@
-package moment.moment.application;
+package moment.moment.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
